MA-21895 Add WITH_CUSTOMIZED_SECURE_UI Marco
This macro is only used on imx8mq. It is false by default to show full
-screen confirmationUI.
If this macro is equal to true, then the position and size of the TUI
need to be modified.
And the DCSS driver will update the scaler and dtg
registers based on this macro.

Test: Display secure UI on 4K and 1080P screen.

Change-Id: Idce0f7e9351cecc2eac066d6a28256b3474067c2
Signed-off-by: He Zhai <zhai.he@nxp.com>
Reviewed-on: http://androidsource.nxp.com/project/23397
Reviewed-by: Elven Wang <elven.wang@nxp.com>